// src/lib/types/vue3/runtime-core.d.ts
export declare interface AppContext {
  app: App
  config: AppConfig
  mixins: ComponentOptions[]
  components: Record<string, Component>
  directives: Record<string, Directive>
  provides: Record<string | symbol, any>
  /* Excluded from this release type: optionsCache */
  /* Excluded from this release type: propsCache */
  /* Excluded from this release type: emitsCache */
  /* Excluded from this release type: reload */
  /* Excluded from this release type: filters */
}
// ...
export declare interface ComponentInternalInstance {
  uid: number
  type: ConcreteComponent
  parent: ComponentInternalInstance | null
  root: ComponentInternalInstance
  appContext: AppContext
  /**
   * Vnode representing this component in its parent's vdom tree
   */
  vnode: VNode
  /* Excluded from this release type: next */
  /**
   * Root vnode of this component's own vdom tree
   */
  subTree: VNode
  /**
   * Render effect instance
   */
  effect: ReactiveEffect
  /**
   * Bound effect runner to be passed to schedulers
   */
  update: SchedulerJob
  /* Excluded from this release type: render */
  /* Excluded from this release type: ssrRender */
  /* Excluded from this release type: provides */
  /* Excluded from this release type: scope */
  /* Excluded from this release type: accessCache */
  /* Excluded from this release type: renderCache */
  /* Excluded from this release type: components */
  /* Excluded from this release type: directives */
  /* Excluded from this release type: filters */
  /* Excluded from this release type: propsOptions */
  /* Excluded from this release type: emitsOptions */
  /* Excluded from this release type: inheritAttrs */
  /**
   * is custom element?
   */
  isCE?: boolean
  /**
   * custom element specific HMR method
   */
  ceReload?: (newStyles?: string[]) => void
  proxy: ComponentPublicInstance | null
  exposed: Record<string, any> | null
  exposeProxy: Record<string, any> | null
  /* Excluded from this release type: withProxy */
  /* Excluded from this release type: ctx */
  data: Data
  props: Data
  attrs: Data
  slots: InternalSlots
  refs: Data
  emit: EmitFn
  /* Excluded from this release type: emitted */
  /* Excluded from this release type: propsDefaults */
  /* Excluded from this release type: setupState */
  /* Excluded from this release type: devtoolsRawSetupState */
  /* Excluded from this release type: setupContext */
  /* Excluded from this release type: suspense */
  /* Excluded from this release type: suspenseId */
  /* Excluded from this release type: asyncDep */
  /* Excluded from this release type: asyncResolved */
  isMounted: boolean
  isUnmounted: boolean
  isDeactivated: boolean
  /* Excluded from this release type: bc */
  /* Excluded from this release type: c */
  /* Excluded from this release type: bm */
  /* Excluded from this release type: m */
  /* Excluded from this release type: bu */
  /* Excluded from this release type: u */
  /* Excluded from this release type: bum */
  /* Excluded from this release type: um */
  /* Excluded from this release type: rtc */
  /* Excluded from this release type: rtg */
  /* Excluded from this release type: a */
  /* Excluded from this release type: da */
  /* Excluded from this release type: ec */
  /* Excluded from this release type: sp */
}
// ...
export declare interface ComponentOptionsBase<
  Props,
  RawBindings,
  D,
  C extends ComputedOptions,
  M extends MethodOptions,
  Mixin extends ComponentOptionsMixin,
  Extends extends ComponentOptionsMixin,
  E extends EmitsOptions,
  EE extends string = string,
  Defaults = {}
> extends LegacyOptions<Props, D, C, M, Mixin, Extends>,
    ComponentInternalOptions,
    ComponentCustomOptions {
  setup?: (
    this: void,
    props: Readonly<
      LooseRequired<
        Props &
          UnionToIntersection<ExtractOptionProp<Mixin>> &
          UnionToIntersection<ExtractOptionProp<Extends>>
      >
    >,
    ctx: SetupContext<E>
  ) => Promise<RawBindings> | RawBindings | RenderFunction | void
  name?: string
  template?: string | object
  render?: Function
  components?: Record<string, Component>
  directives?: Record<string, Directive>
  inheritAttrs?: boolean
  emits?: (E | EE[]) & ThisType<void>
  expose?: string[]
  serverPrefetch?(): Promise<any>
  compilerOptions?: RuntimeCompilerOptions
  /* Excluded from this release type: ssrRender */
  /* Excluded from this release type: __ssrInlineRender */
  /* Excluded from this release type: __asyncLoader */
  /* Excluded from this release type: __asyncResolved */
  call?: (this: unknown, ...args: unknown[]) => never
  __isFragment?: never
  __isTeleport?: never
  __isSuspense?: never
  __defaults?: Defaults
}
// ...
export declare type ComponentPublicInstance<
  P = {}, // props type extracted from props option
  B = {}, // raw bindings returned from setup()
  D = {}, // return from data()
  C extends ComputedOptions = {},
  M extends MethodOptions = {},
  E extends EmitsOptions = {},
  PublicProps = P,
  Defaults = {},
  MakeDefaultsOptional extends boolean = false,
  Options = ComponentOptionsBase<any, any, any, any, any, any, any, any, any>
> = {
  $: ComponentInternalInstance
  $data: D
  $props: MakeDefaultsOptional extends true
    ? Partial<Defaults> & Omit<P & PublicProps, keyof Defaults>
    : P & PublicProps
  $attrs: Data
  $refs: Data
  $slots: Slots
  $root: ComponentPublicInstance | null
  $parent: ComponentPublicInstance | null
  $emit: EmitFn<E>
  $el: any
  $options: Options & MergedComponentOptionsOverride
  $forceUpdate: () => void
  $nextTick: typeof nextTick
  $watch(source: string | Function, cb: Function, options?: WatchOptions): WatchStopHandle
} & P &
  ShallowUnwrapRef<B> &
  UnwrapNestedRefs<D> &
  ExtractComputedReturns<C> &
  M &
  ComponentCustomProperties
// ...
export declare type ComputedOptions = Record<
  string,
  ComputedGetter<any> | WritableComputedOptions<any>
>
// ...
declare type Data = Record<string, unknown>
// ...
export declare interface DirectiveBinding<V = any> {
  instance: ComponentPublicInstance | null
  value: V
  oldValue: V | null
  arg?: string
  modifiers: DirectiveModifiers
  dir: ObjectDirective<any, V>
}

export declare type DirectiveHook<T = any, Prev = VNode<any, T> | null, V = any> = (
  el: T,
  binding: DirectiveBinding<V>,
  vnode: VNode<any, T>,
  prevVNode: Prev
) => void

declare type DirectiveModifiers = Record<string, boolean>

export { effect }

export { EffectScheduler }

export { EffectScope }

export { effectScope }

declare type EmitFn<
  Options = ObjectEmitsOptions,
  Event extends keyof Options = keyof Options
> = Options extends Array<infer V>
  ? (event: V, ...args: any[]) => void
  : {} extends Options
  ? (event: string, ...args: any[]) => void
  : UnionToIntersection<
      {
        [key in Event]: Options[key] extends (...args: infer Args) => any
          ? (event: key, ...args: Args) => void
          : (event: key, ...args: any[]) => void
      }[Event]
    >

export declare type EmitsOptions = ObjectEmitsOptions | string[]
// ...
declare type MergedComponentOptionsOverride = {
  beforeCreate?: MergedHook
  created?: MergedHook
  beforeMount?: MergedHook
  mounted?: MergedHook
  beforeUpdate?: MergedHook
  updated?: MergedHook
  activated?: MergedHook
  deactivated?: MergedHook
  /** @deprecated use `beforeUnmount` instead */
  beforeDestroy?: MergedHook
  beforeUnmount?: MergedHook
  /** @deprecated use `unmounted` instead */
  destroyed?: MergedHook
  unmounted?: MergedHook
  renderTracked?: MergedHook<DebuggerHook>
  renderTriggered?: MergedHook<DebuggerHook>
  errorCaptured?: MergedHook<ErrorCapturedHook>
}

/* Excluded from this release type: mergeDefaults */

declare type MergedHook<T = () => void> = T | T[]

export declare function mergeProps(...args: (Data & VNodeProps)[]): Data

export declare interface MethodOptions {
  [key: string]: Function
}
// ...
export declare interface ObjectDirective<T = any, V = any> {
  created?: DirectiveHook<T, null, V>
  beforeMount?: DirectiveHook<T, null, V>
  mounted?: DirectiveHook<T, null, V>
  beforeUpdate?: DirectiveHook<T, VNode<any, T>, V>
  updated?: DirectiveHook<T, VNode<any, T>, V>
  beforeUnmount?: DirectiveHook<T, null, V>
  unmounted?: DirectiveHook<T, null, V>
  getSSRProps?: SSRDirectiveHook
  deep?: boolean
}
// ...
declare type PublicProps = VNodeProps & AllowedComponentProps & ComponentCustomProps
// ...
export declare interface RendererElement extends RendererNode {}

declare interface RendererInternals<HostNode = RendererNode, HostElement = RendererElement> {
  p: PatchFn
  um: UnmountFn
  r: RemoveFn
  m: MoveFn
  mt: MountComponentFn
  mc: MountChildrenFn
  pc: PatchChildrenFn
  pbc: PatchBlockChildrenFn
  n: NextFn
  o: RendererOptions<HostNode, HostElement>
}

export declare interface RendererNode {
  [key: string]: any
}
// ...
export declare type Slots = Readonly<InternalSlots>

/**
 * Use this for features where legacy usage is still possible, but will likely
 * lead to runtime error if compat is disabled. (warn in all cases)
 */
declare function softAssertCompatEnabled(
  key: DeprecationTypes,
  instance: ComponentInternalInstance | null,
  ...args: any[]
): boolean

export declare const ssrContextKey: unique symbol

declare type SSRDirectiveHook = (binding: DirectiveBinding, vnode: VNode) => Data | undefined

/* Excluded from this release type: ssrUtils */

export declare const Static: unique symbol

export { stop_2 as stop }

export declare const Suspense: {
  new (): {
    $props: VNodeProps & SuspenseProps
  }
  __isSuspense: true
}

export declare interface SuspenseBoundary {
  vnode: VNode<RendererNode, RendererElement, SuspenseProps>
  parent: SuspenseBoundary | null
  parentComponent: ComponentInternalInstance | null
  isSVG: boolean
  container: RendererElement
  hiddenContainer: RendererElement
  anchor: RendererNode | null
  activeBranch: VNode | null
  pendingBranch: VNode | null
  deps: number
  pendingId: number
  timeout: number
  isInFallback: boolean
  isHydrating: boolean
  isUnmounted: boolean
  effects: Function[]
  resolve(force?: boolean): void
  fallback(fallbackVNode: VNode): void
  move(container: RendererElement, anchor: RendererNode | null, type: MoveType): void
  next(): RendererNode | null
  registerDep(instance: ComponentInternalInstance, setupRenderEffect: SetupRenderEffectFn): void
  unmount(parentSuspense: SuspenseBoundary | null, doRemove?: boolean): void
}
// ...
export declare interface TransitionHooks<HostElement extends RendererElement = RendererElement> {
  mode: BaseTransitionProps['mode']
  persisted: boolean
  beforeEnter(el: HostElement): void
  enter(el: HostElement): void
  leave(el: HostElement, remove: () => void): void
  clone(vnode: VNode): TransitionHooks<HostElement>
  afterLeave?(): void
  delayLeave?(el: HostElement, earlyRemove: () => void, delayedLeave: () => void): void
  delayedLeave?(): void
}
// ...
export declare interface VNode<
  HostNode = RendererNode,
  HostElement = RendererElement,
  ExtraProps = {
    [key: string]: any
  }
> {
  /* Excluded from this release type: __v_isVNode */
  /* Excluded from this release type: __v_skip */
  type: VNodeTypes
  props: (VNodeProps & ExtraProps) | null
  key: string | number | symbol | null
  ref: VNodeNormalizedRef | null
  /**
   * SFC only. This is assigned on vnode creation using currentScopeId
   * which is set alongside currentRenderingInstance.
   */
  scopeId: string | null
  /* Excluded from this release type: slotScopeIds */
  children: VNodeNormalizedChildren
  component: ComponentInternalInstance | null
  dirs: DirectiveBinding[] | null
  transition: TransitionHooks<HostElement> | null
  el: HostNode | null
  anchor: HostNode | null
  target: HostElement | null
  targetAnchor: HostNode | null
  /* Excluded from this release type: staticCount */
  suspense: SuspenseBoundary | null
  /* Excluded from this release type: ssContent */
  /* Excluded from this release type: ssFallback */
  shapeFlag: number
  patchFlag: number
  /* Excluded from this release type: dynamicProps */
  /* Excluded from this release type: dynamicChildren */
  appContext: AppContext | null
  /* Excluded from this release type: memo */
  /* Excluded from this release type: isCompatRoot */
  /* Excluded from this release type: ce */
}
// ...
export declare type VNodeNormalizedChildren = string | VNodeArrayChildren | RawSlots | null

declare type VNodeNormalizedRef = VNodeNormalizedRefAtom | VNodeNormalizedRefAtom[]

declare type VNodeNormalizedRefAtom = {
  i: ComponentInternalInstance
  r: VNodeRef
  k?: string
  f?: boolean
}

export declare type VNodeProps = {
  key?: string | number | symbol
  ref?: VNodeRef
  ref_for?: boolean
  ref_key?: string
  onVnodeBeforeMount?: VNodeMountHook | VNodeMountHook[]
  onVnodeMounted?: VNodeMountHook | VNodeMountHook[]
  onVnodeBeforeUpdate?: VNodeUpdateHook | VNodeUpdateHook[]
  onVnodeUpdated?: VNodeUpdateHook | VNodeUpdateHook[]
  onVnodeBeforeUnmount?: VNodeMountHook | VNodeMountHook[]
  onVnodeUnmounted?: VNodeMountHook | VNodeMountHook[]
}

declare type VNodeRef = string | Ref | ((ref: object | null, refs: Record<string, any>) => void)

export declare type VNodeTypes =
  | string
  | VNode
  | Component
  | typeof Text_2
  | typeof Static
  | typeof Comment_2
  | typeof Fragment
  | typeof TeleportImpl
  | typeof SuspenseImpl
// ...
export declare interface WatchOptions<Immediate = boolean> extends WatchOptionsBase {
  immediate?: Immediate
  deep?: boolean
}